requirements.txt
streamlit>=1.32
psycopg2-binary>=2.9
pandas>=2.0
plotly>=5.20
streamlit-authenticator>=0.3
openpyxl>=3.1
python-dotenv>=1.0
bcrypt>=4.0
